third_party/blink/web_tests/virtual/lna-allow-potentially-trustworthy-same-origin/external/wpt/fetch/private-network-access/fetch-from-treat-as-public.https.window-expected.txt

This is a testharness.js-based test.
FAIL treat-as-public-address to local: failed preflight. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS treat-as-public-address to local: success.
PASS treat-as-public-address to private: failed preflight.
PASS treat-as-public-address to private: success.
PASS treat-as-public-address to public: no preflight required.
Harness: the test ran to completion.
